[Flex] crossAxisDirection should take directionality into consideration https://bugs.webkit.org/show_bug.cgi?id=296402 rdar://156540996
Reviewed by Alan Baradlay.

The goal of crossAxisDirection, as the name implies, is to map the
direction in which the flexbox's cross axis flows to a physical direction.
It correctly looks at flex-wrap since flex-wrap: wrap-reverse can flip
the direction of the cross axis.

However, there is an issue when the cross axis is in the same axis as
the flexbox's inline axis. In this case, the start and end of the inline
axis can be affected by the direction property. In those cases, we need
to look at which way the inline direction flows in order to determine
whether the cross axis is simply that direction or needs to be flipped
in the case of flex-wrap: wrap-reverse. This can be done by taking
writingMode().inlineDirection() instead of writingMode().blockDirection()
when the cross axis is in the inline axis and passing it into the switch
statement.

I also decided to move this function definition from the header to the
source since it is not clear to me why I decided to make it inline as I
do not think there is an obvious benefit that outweighs the - albeit
likely small - build time increase.

Canonical link: https://commits.webkit.org/297876@main