Add support for HTML 'title' attribute (tooltips) on Catalyst https://bugs.webkit.org/show_bug.cgi?id=296470 rdar://79416560
Reviewed by Tim Horton.

Added support for tooltips on Catalyst.

This includes:
- Linking UIKitMacHelper
- Declaring needed protocols and interfaces
- Implementing toolTipChanged

The bulk of the functionality is in -[WKContentView _toolTipChanged:],
which creates a UIToolTipInteraction if it doesn't exist. If a
UIToolTipInteraction exists, we update the UIToolTipInteraction with
the tooltip's new string.

Notice, _toolTipChanged's has a windowChangedKeyState call and uses a
UINSSharedApplicationDelegate. This is for updating the tooltip
after the initial creation. Implementing UIToolTipInteractionDelegate's
toolTipInteraction:configurationAtPoint: is not enough to update the tooltip
because...
1. UIView's addInteraction and removeInteraction are only for separate views.
That is, if there are tooltips A and B on the same view, whichever tooltip is
triggered first will stay no matter where you hover within the view. The interaction is not
removed until you move your mouse to a different view.
2. There is no API to update (text or show/hide) the tooltip while it is displayed.

To work around #2, we need to indirectly trigger AppKit's _displayToolTipIfNecessaryIgnoringTime,
which is the logic to update the tooltip, but through UIKit.

See below for more details.

* Source/WTF/wtf/PlatformHave.h:
* Source/WebKit/Configurations/WebKit.xcconfig:
* Source/WebKit/Platform/spi/ios/UIKitSPI.h:

These files above have necessary declarations that are used for working around #2,
adding a HAVE macro, and linking UIKitMacHelper.

* Source/WebKit/UIProcess/ios/PageClientImplIOS.mm:
(WebKit::PageClientImpl::toolTipChanged):

Calls -[WKContentView _toolTipChanged:], which contains the logic
for supporting tooltips.

* Source/WebKit/UIProcess/ios/WKContentViewInteraction.h:
* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView _toolTipChanged:]):

This is where the support is implemented. To circumvent #2, we first use
UINSSharedApplicationDelegate, which is a UINSApplicationDelegate, then obtaining
a UINSWindow. UINSWindow has a UINSSceneView that has a NSViewDynamicToolTipManager
property. The NSViewDynamicToolTipManager can be used to call windowChangedKeyState.
We call windowChangedKeyState to trigger a chain of function calls that will eventually
trigger the update tooltip function. The chain is as follows: windowChangedKeyState calls
_restartMovementTracking which calls _displayToolTipIfNecessaryIgnoringTime, which updates
the tooltip.

(-[WKContentView toolTipInteraction:configurationAtPoint:]):

Canonical link: https://commits.webkit.org/297878@main