Very low framerate on youtube.com when system is under load https://bugs.webkit.org/show_bug.cgi?id=296474 rdar://156331158
Reviewed by Jer Noble.

We enqueued samples to be decoded and rendered using a standard priority thread
to be decoded by the WebCoreDecompressionSession.
When the AVSampleBufferDisplayLayer was doing this work, it used a queue
with elevated priority to decode video. This allowed videos to play smoothly
even when the system was under very heavy load.

We make the VideoMediaSampleRenderer use a common queue now with UserInteractive
priority.
We also allow for a way to optionally provide a workqueue to the WebCoreDecompressionSession
to enqueue frame to the VTDecompressionSession. We now share the VideoMediaSampleRenderer's
queue with the WebCoreDecompressionSession so that both tasks run at the same priority.

No change to JS observable behaviour, covered by existing tests.

Canonical link: https://commits.webkit.org/297881@main