[Absolute Positioning][Grid] Expand alignment within static position rectangle to align-self center for simple content https://bugs.webkit.org/show_bug.cgi?id=296455 rdar://156643206
Reviewed by Alan Baradlay.

Continues to build upon alignment within the static position rectangle
for abspos children of grids. We keep the same constraints as we
curently have (horizontal-tb LTR content) for end alignment and just
expand eligibility for center. We continue to compute the offset to
satisfy the alignment via resolveAlignmentShift.

* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-self-center-large-border-padding.html: Added.
Grid with a large border and padding on the top aht bottom. Content box
size is fairly large to make it obvious if the item is not being aligned
properly in the center.

* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-self-center.html: Added.
Simple testcase to make sure alignment works in a basic scenario.

Canonical link: https://commits.webkit.org/297900@main