[GStreamer][WebRTC] Handle H.264 profile-level-id caps field in outgoing video sources https://bugs.webkit.org/show_bug.cgi?id=296263
Reviewed by Xabier Rodriguez-Calvar.

If the profile=level-id attribute is present, use it for configuring the encoder and remove it from
the caps, otherwise the caps negotiation with the payloader might fail.

Driving-by, make sure the CSF0 bitfield is set on the constraint flags for the constrained-baseline
profile.

* Source/WebCore/platform/mediastream/gstreamer/GStreamerVideoRTPPacketizer.cpp:
(WebCore::GStreamerVideoRTPPacketizer::create):

Canonical link: https://commits.webkit.org/297913@main