[ macOS ] media/media-source/media-source-rvfc-paused-offscreen.html is a flakey text failure https://bugs.webkit.org/show_bug.cgi?id=296334 rdar://156415981
Reviewed by Youenn Fablet.

The test exposed two issues:
1- If the seeked event was fired after the requestVideoFrameCallback resolved,
   it would have lead to an expectation failure due to the ordering of messages varying.
2- From time to time, the rVFC would be resolved for the same frame (here the first) even though time progressed.

Issue 1 is resolved by amending the test so that we wait for the seeked event
before checking the rVFC result.

Issue 2 occurred because whenever we set the TrackBuffer's needsReenqueueing flag
we would only ever clear it if we did actually need to re-enqueue.
As such, on playback startup, or after a seek, samples were always unnecessarily
enqueued and decoded twice.
This regression was introduced a long time ago in 251376@main.

As the rVFC is resolved whenever we have a new frame to display, when
decoding the same frame a second time and it was still in time for display
the rVFC would be resolved with that same frame again.

The TrackBuffer shouldn't be managing its state and how frames should be enqueued to it,
the SourceBufferPrivate should and treat the TrackBuffer as a container only.
We will address this issue in a follow-up change.

* LayoutTests/media/media-source/media-source-rvfc-paused-offscreen.html:
* LayoutTests/media/media-source/media-source-rvfc-paused.html:
* LayoutTests/platform/mac/TestExpectations:
* Source/WebCore/platform/graphics/TrackBuffer.cpp:
(WebCore::TrackBuffer::reenqueueMediaForTime): Clear the needsReenqueueing flag as soon
as the method is called: there's nothing to re-enqueue in the future.

Canonical link: https://commits.webkit.org/297924@main