[GTK][WPE] Rename DMABufRendererBufferFormat as RendererBufferFormat https://bugs.webkit.org/show_bug.cgi?id=296481
Reviewed by Carlos Garcia Campos.

The concept of a buffer format can apply to other types of buffers other
than DMA-BUF ones, therefore remove the DMABufRendererBufferFormat,
leaving it as plainly RendererBufferFormat.

A class named RendererBufferFormat already existed, but its name was
confusing because it is used to describe the properties of buffers being
used for rendering in order to display the information in webkit://gpu.
Therefore, its name gets changed to RendererBufferDescription, which
is clearer given its intended usage.

Canonical link: https://commits.webkit.org/297929@main