[WPE] Convert AcceleratedSurfaceLibWPE into a RenderTarget of AcceleratedSurfaceDMABuf https://bugs.webkit.org/show_bug.cgi?id=296561
Reviewed by Adrian Perez de Castro.

This is in preparation to rename AcceleratedSurfaceDMABuf as
AcceleratedSurface and use different RenderTarget implementation for the
different buffer types. RenderTargetWPEBackend is a special case of
RenderTarget that only allows one target that is resized instead of
re-created and has its own way to communicate with the UI process. For
the other targets that use WebKit IPC messages for the communication a
base class RenderTargetShareableBuffer has been added,

Canonical link: https://commits.webkit.org/298029@main