[Skia] Incorrect font rendering of composed characters with skia and certain fonts https://bugs.webkit.org/show_bug.cgi?id=291272
Reviewed by Adrian Perez de Castro.

The problem happens with some sequences like "R̥g"
which is composed of characters (#x0052 #x0325)
LATIN CAPITAL LETTER R with COMBINING RING BELOW. Since the system
fallback only allows to check for a single character, if we find a font
that contains a glyph for 'R' but that can't render the combining
sequence we end up falling back to the last resort fallback font, which
depending in the system might not be able to render the sequence either.
So, when we fail to find a font that can render the sequence, we should
try to find another font in the system that contains all other
characters in the sequence as we do with cairo. But Skia API doesn't
allow to query a set of characters, we would need to call the
matchFamilyStyleCharacter() for every character which would be very
slow. We need to use fontconfig API for that, looking for a font that
contains all characters in the sequence or at least the one having more
charcter coverage. We can take advantage that we have to use fontconfig
API directly for system fallbacks to being back the FontSet cache that
will make it even faster. See https://bugs.webkit.org/show_bug.cgi?id=203544

Canonical link: https://commits.webkit.org/298033@main