No loading spinner during APMP playback after scrubbing and waiting for the track to play in Safari https://bugs.webkit.org/show_bug.cgi?id=296667 rdar://156606910
Reviewed by Jer Noble and Eric Carlson.

For the spinner to show while in the LMK player, isLoadingPublisher has to be true
on Playable surface.

Its value was always set to NO. We now set it to YES whenever playback has
stalled, which includes when loading data after a seek.

Manually verified as working as intended.

Canonical link: https://commits.webkit.org/298054@main