WebVTT test assumes default line alignment is "center", when it should be "start" https://bugs.webkit.org/show_bug.cgi?id=296706 rdar://156715310
Reviewed by Eric Carlson.

In /webvtt/rendering/cues-with-video/processing-model/line_50_percent-ref.html, the
<span class="cue"> is given margin-top: -4.5px. This is put there to render the cue
exactly vertically in the middle of the video. However, this is not the correct position.
According to section 3.3 of the WebVTT spec under "A line alignment", by default the
line alignment is set to start (which is the top edge for horizontal cues). Based on
this, I do not think the cue is supposed to be exactly centered, but rather the top edge
should be positioned at the center.

To fix this, we should simply remove margin-top: -4.5px from line_50_percent-ref.html and
line_50_percent-expected.html.

* LayoutTests/imported/w3c/web-platform-tests/webvtt/rendering/cues-with-video/processing-model/line_50_percent-expected.html:
* LayoutTests/imported/w3c/web-platform-tests/webvtt/rendering/cues-with-video/processing-model/line_50_percent-ref.html:

Canonical link: https://commits.webkit.org/298086@main