Add quirk to allow downloads on dropbox.com https://bugs.webkit.org/show_bug.cgi?id=296749 rdar://155649992
Reviewed by Brent Fulgham.

https://www.dropbox.com/download?os=mac has a content-security-policy sandbox that does not
allow downloads, but only if requested with a Safari user-agent.  This was not an issue until
we increased our standards compliance and interoperability and security in 283667@main by
honoring content-security-policy sandboxes with regards to downloads initiated by the main frame.
Unfortunately, this broke the ability to download the dropbox installer, which dropbox.com is
saying should not be allowed on its website.  Add a quirk to get the old behavior until they fix it.

* Source/WebCore/loader/DocumentLoader.cpp:
(WebCore::DocumentLoader::continueAfterContentPolicy):
* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldAllowDownloadsInSpiteOfCSP const):
* Source/WebCore/page/Quirks.h:

Canonical link: https://commits.webkit.org/298093@main