Change LineDirectionMode to enum class https://bugs.webkit.org/show_bug.cgi?id=296775 rdar://157254054
Reviewed by Vitor Roriz.

Instead of a plain enum let's make it a strongly typed enum. Also rename
it from LineDirectionMode to LineDirection and values of
Horizontal/Vertical instead of HorizontalLine/VerticalLine. This is
primarily used to know the direction of the line for the purposes of
baseline alignment so I don't think "mode," really adds much.

Canonical link: https://commits.webkit.org/298128@main