[Swift in WebKit] Fix the access levels of some JSC private headers to properly be project-level https://bugs.webkit.org/show_bug.cgi?id=296752 rdar://157225138
Reviewed by Abrar Rahman Protyasha and Keith Miller.

There are a few headers in JSC that are private but include project-level headers. This is invalid,
as a header file can only access other header files that are as or more visible than themselves.

This isn't currently a compilation error because these headers aren't actually included from anywhere
outside JSC so it never hits the issue. However, with module verification and a private JSC modulemap,
this becomes an issue because module verification is strict about correctness and validity.

Fix by simply making the affected private headers project-level.

* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/298138@main