SharedBufferBuilder::take() should return a contiguous SharedBuffer if it only contains a single DataSegment. https://bugs.webkit.org/show_bug.cgi?id=296684 rdar://157093750
Reviewed by Youenn Fablet.

Make SharedBufferBuilder always return a SharedBuffer if the data it holds
is contiguous, and a FragmentedSharedBuffer otherwise.
Following this change, you can no longer create an object of the base class
FragmentedSharedBuffer if it is contiguous: it will always be a SharedBuffer
instead.
A side benefits is that it allows to remove a memory allocation whenever
we want to convert a FragmentedSharedBuffer into a SharedBuffer such as when
when using a SharedBufferBuilder and appending a single segment to it:
SharedBufferBuilder::takeAsContiguous() is now a no-op if the underlying
buffer is already contiguous.

API tests added.

* Source/WTF/wtf/TypeTraits.h: Adding RemoveCVSmartPointer convenience template.
* Source/WebCore/Modules/highlight/AppHighlight.h:
(IPC::AsyncReplyError<WebCore::AppHighlight>::create): Remove use of FragmentedSharedBuffer
that was in effect always contiguous.
* Source/WebCore/Modules/highlight/AppHighlightRangeData.cpp:
(WebCore::AppHighlightRangeData::create): Ditto
(WebCore::AppHighlightRangeData::toSharedBuffer const):
* Source/WebCore/Modules/highlight/AppHighlightRangeData.h:
* Source/WebCore/Modules/highlight/AppHighlightStorage.cpp: Ditto
(WebCore::AppHighlightStorage::restoreAndScrollToAppHighlight):
* Source/WebCore/Modules/highlight/AppHighlightStorage.h: Ditto

Canonical link: https://commits.webkit.org/298168@main