[Layout][Grid] Mitigate incorrect baseline for vertical-lr flex/grid boxes. https://bugs.webkit.org/show_bug.cgi?id=296443 <rdar://156628380>
Reviewed by Alan Baradlay.

Inline layout currently performs an unnecessary coordinate flip in

InlineDisplayContentBuilder::mapInlineRectLogicalToVisual()

for vertical-lr content.

This is not necessary for boxes that provide the correct coordinates
and baselines, which are then broken by this flip.

This change introduces a temporary fix in

LayoutIntegrationGeometryUpdate::baselineForBox()

by performing another flip for line-inverted writing modes. This ensures
that the baselines computed by the LayoutIntegrationGeometryUpdate will
allow the content to be rendered correctly.

 * LayoutTests/TestExpectations:
 * LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-container-baseline-001-expected.txt:
 * Source/WebCore/layout/integration/LayoutIntegrationBoxGeometryUpdater.cpp:
 (WebCore::LayoutIntegration::baselineForBox):

Canonical link: https://commits.webkit.org/298204@main