Don't create fragmented blob made of other blobs with fetch data https://bugs.webkit.org/show_bug.cgi?id=296531 rdar://156827880
Reviewed by Youenn Fablet.

Follow-up on 297885@main. Instead of constructing a Blob
made of multiple blobs, which would create a lot of Blob URLs, we instead
add the ability to create a Blob from a FragmentedSharedBuffer. A BlobPart
is also amended to allow for holding a SharedBuffer.

To reduce the amount of elements in the FragmentedSharedBuffer, we ensure
that the DataSegments it contains are at least 8MiB in size.

No change in JS observable behaviours. Covered by existing tests.

* Source/WebCore/platform/network/BlobPart.h:
(WebCore::BlobPart::BlobPart): Add SharedBuffer as variant, clean-up
code to follow SaferCPP coding guidelines.
(WebCore::BlobPart::takeData): Renamed for consistency and better indicate how the code was used.
(WebCore::BlobPart::moveData): Deleted.

Canonical link: https://commits.webkit.org/298227@main